create libsend for outgoing mail
