+++ Makefile.autosetup
@@ -73,7 +73,7 @@ NEOMUTTOBJS=	browser.o commands.o command_parse.o \
 		mutt_parse.o mutt_signal.o mutt_socket.o mutt_thread.o mx.o \
 		myvar.o opcodes.o pager.o pattern.o postpone.o progress.o \
 		recvattach.o recvcmd.o resize.o rfc3676.o score.o \
-		send.o sendlib.o sidebar.o smtp.o sort.o state.o status.o \
+		sidebar.o sort.o state.o status.o \
 		system.o version.o
 
 @if !HAVE_WCSCASECMP
@@ -468,6 +468,19 @@ $(PWD)/pop:
 	$(MKDIR_P) $(PWD)/pop
 
 ###############################################################################
+# libsend
+LIBSEND=	libsend.a
+LIBSENDOBJS=	send/body.o send/config.o send/header.o send/multipart.o send/send.o send/sendlib.o send/sendmail.o send/smtp.o
+CLEANFILES+=	$(LIBSEND) $(LIBSENDOBJS)
+ALLOBJS+=	$(LIBSENDOBJS)
+
+$(LIBSEND): $(PWD)/send $(LIBSENDOBJS)
+	$(AR) cr $@ $(LIBSENDOBJS)
+	$(RANLIB) $@
+$(PWD)/send:
+	$(MKDIR_P) $(PWD)/send
+
+###############################################################################
 # libstore
 @if HAVE_BDB
 LIBSTOREOBJS+=	store/bdb.o
@@ -532,8 +545,9 @@ $(ALLOBJS):
 MUTTLIBS+=	$(LIBAUTOCRYPT) $(LIBPOP) $(LIBNNTP) $(LIBCOMPMBOX) \
 		$(LIBSTORE) $(LIBGUI) $(LIBDEBUG) $(LIBMBOX) $(LIBNOTMUCH) \
 		$(LIBMAILDIR) $(LIBNCRYPT) $(LIBIMAP) $(LIBCONN) $(LIBHCACHE) \
-		$(LIBCOMPRESS) $(LIBBCACHE) $(LIBHISTORY) $(LIBALIAS) $(LIBCORE) \
-		$(LIBCONFIG) $(LIBEMAIL) $(LIBADDRESS) $(LIBMUTT)
+		$(LIBCOMPRESS) $(LIBBCACHE) $(LIBHISTORY) $(LIBALIAS) \
+		$(LIBSEND) $(LIBCORE) $(LIBCONFIG) $(LIBEMAIL) $(LIBADDRESS) \
+		$(LIBMUTT)
 
 # neomutt
 $(NEOMUTT): $(GENERATED) $(NEOMUTTOBJS) $(MUTTLIBS)
