+++ commands.c
@@ -64,9 +64,9 @@
 #include "options.h"
 #include "pager.h"
 #include "protos.h"
-#include "sendlib.h"
 #include "sort.h"
 #include "ncrypt/lib.h"
+#include "send/lib.h"
 #ifdef USE_IMAP
 #include "imap/lib.h"
 #endif
