+++ mutt_attach.c
@@ -52,9 +52,9 @@
 #include "options.h"
 #include "pager.h"
 #include "protos.h"
-#include "sendlib.h"
 #include "state.h"
 #include "ncrypt/lib.h"
+#include "send/lib.h"
 #ifdef USE_IMAP
 #include "imap/lib.h"
 #endif
