+++ mutt_globals.h
@@ -75,7 +75,6 @@ WHERE struct RegexList Alternates INITVAL(STAILQ_HEAD_INITIALIZER(Alternates)); 
 WHERE struct RegexList UnAlternates INITVAL(STAILQ_HEAD_INITIALIZER(UnAlternates));           ///< List of regexes to blacklist false matches in Alternates
 
 /* flags for received signals */
-WHERE SIG_ATOMIC_VOLATILE_T SigAlrm;  ///< true after SIGALRM is received
 WHERE SIG_ATOMIC_VOLATILE_T SigInt;   ///< true after SIGINT is received
 WHERE SIG_ATOMIC_VOLATILE_T SigWinch; ///< true after SIGWINCH is received
 
@@ -104,8 +103,6 @@ WHERE char *C_AutocryptDir;                  ///< Config: Location of autocrypt 
 WHERE char *C_ConfigCharset;                 ///< Config: Character set that the config files are in
 WHERE char *C_CryptProtectedHeadersSubject;  ///< Config: Use this as the subject for encrypted emails
 WHERE char *C_DateFormat;                    ///< Config: strftime format string for the `%d` expando
-WHERE char *C_DsnNotify;                     ///< Config: Request notification for message delivery or delay
-WHERE char *C_DsnReturn;                     ///< Config: What to send as a notification of message delivery or delay
 WHERE char *C_Editor;                        ///< Config: External command to use as an email editor
 WHERE char *C_ExternalSearchCommand;         ///< Config: External search command
 WHERE char *C_Hostname;                      ///< Config: Fully-qualified domain name of this machine
@@ -134,9 +131,6 @@ WHERE char *C_NewMailCommand;                ///< Config: External command to ru
 WHERE char *C_Realname;                      ///< Config: Real name of the user
 WHERE char *C_Shell;                         ///< Config: External command to run subshells in
 WHERE char *C_SimpleSearch;                  ///< Config: Pattern to search for when search doesn't contain ~'s
-#ifdef USE_SMTP
-WHERE char *C_SmtpUrl;                       ///< Config: (smtp) Url of the SMTP server
-#endif
 WHERE char *C_Spoolfile;                     ///< Config: Inbox
 WHERE char *C_StatusFormat;                  ///< Config: printf-like format string for the index's status line
 WHERE char *C_TsStatusFormat;                ///< Config: printf-like format string for the terminal's status (window title)
