+++ postpone.c
@@ -55,11 +55,10 @@
 #include "options.h"
 #include "protos.h"
 #include "rfc3676.h"
-#include "send.h"
-#include "sendlib.h"
 #include "sort.h"
 #include "state.h"
 #include "ncrypt/lib.h"
+#include "send/lib.h"
 #ifdef USE_IMAP
 #include "imap/lib.h"
 #endif
