+++ recvattach.c
@@ -58,10 +58,9 @@
 #include "opcodes.h"
 #include "options.h"
 #include "recvcmd.h"
-#include "send.h"
-#include "sendlib.h"
 #include "state.h"
 #include "ncrypt/lib.h"
+#include "send/lib.h"
 #ifdef ENABLE_NLS
 #include <libintl.h>
 #endif
