+++ recvcmd.c
@@ -50,9 +50,8 @@
 #include "muttlib.h"
 #include "options.h"
 #include "protos.h"
-#include "send.h"
-#include "sendlib.h"
 #include "state.h"
+#include "send/lib.h"
 #ifdef ENABLE_NLS
 #include <libintl.h>
 #endif
