+++ send/multipart.h
@@ -1,0 +1,33 @@
+/**
+ * @file
+ * Manipulate multipart Emails
+ *
+ * @authors
+ * Copyright (C) 2020 Richard Russon <rich@flatcap.org>
+ *
+ * @copyright
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 2 of the License, or (at your option) any later
+ * version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef MUTT_SEND_MULTIPART_H
+#define MUTT_SEND_MULTIPART_H
+
+struct Body;
+struct ParameterList;
+
+void         mutt_generate_boundary(struct ParameterList *pl);
+struct Body *mutt_make_multipart(struct Body *b);
+struct Body *mutt_remove_multipart(struct Body *b);
+
+#endif /* MUTT_SEND_MULTIPART_H */
